+++ src/client/lib/constants.ts
@@ -10,4 +10,5 @@ export const APP_NAME = 'Pulsarr'
 export const APP_DESCRIPTION = 'Plex watchlist tracker and notification center'
 
 // UI/UX constants
-export const MIN_LOADING_DELAY = 500 // Minimum loading delay for UI feedback (ms)
+export const LOADER_SHOW_DELAY = 125 // Loaders only appear when loading exceeds this (ms), preventing flash on fast responses
+export const MIN_LOADING_DELAY = 500 // Once shown, loaders stay at least this long (ms)
